+++ de.upb.llvm_parser/src-gen/de/upb/llvm_parser/llvm/Vector.java
@@ -2,7 +2,6 @@
  */
 package de.upb.llvm_parser.llvm;
 
-import java.math.BigDecimal;
 
 /**
  * <!-- begin-user-doc -->
@@ -32,12 +31,12 @@ public interface Vector extends Aggregate_Type
    * </p>
    * <!-- end-user-doc -->
    * @return the value of the '<em>Length</em>' attribute.
-   * @see #setLength(BigDecimal)
+   * @see #setLength(int)
    * @see de.upb.llvm_parser.llvm.LlvmPackage#getVector_Length()
    * @model
    * @generated
    */
-  BigDecimal getLength();
+  int getLength();
 
   /**
    * Sets the value of the '{@link de.upb.llvm_parser.llvm.Vector#getLength <em>Length</em>}' attribute.
@@ -47,7 +46,7 @@ public interface Vector extends Aggregate_Type
    * @see #getLength()
    * @generated
    */
-  void setLength(BigDecimal value);
+  void setLength(int value);
 
   /**
    * Returns the value of the '<em><b>Type</b></em>' containment reference.
